requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-cors-headers>=3.8.0
python-dotenv>=0.19.0
numpy>=1.21.0
scikit-learn>=0.24.2
joblib>=1.0.2
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
nltk==3.8.1
google-auth==2.23.4
google-auth-oauthlib<1.1.0,>=0.5
gunicorn==21.2.0
whitenoise==6.6.0
python-social-auth==0.3.6
social-auth-app-django==5.4.0 